requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
django-utils-six==2.0
djangorestframework==3.16.0
drf-yasg==1.21.10
Faker==37.1.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1